slight modularization of server.ts graphql elements have been placed in the 'graphql/' folder for better maintance
