+++ src/server.ts
@@ -1,44 +1,6 @@
-import { ApolloServer, gql } from 'apollo-server';
-
-const typeDefs = gql`
-  type Query {
-    hello: String
-    goodbye: String
-  }
-
-  type User {
-    id: ID
-    name: String
-    email: String
-    birthDate: String
-    cpf: String
-  }
-
-  type LoginResponse {
-    user: User
-    token: Int
-  }
-
-  type Mutation {
-    login(email: String, password: String): LoginResponse
-  }
-`;
-
-const resolvers = {
-  Query: {
-    hello: () => 'Hello world!',
-    goodbye: () => 'Goodbye world!',
-  },
-  Mutation: {
-    login(_, args) {
-      return {
-        user: { id: 123, birthDate: '2020-01-01', email: args.email, cpf: '62693406080' },
-        token: 12,
-      };
-    },
-  },
-};
-
+import { ApolloServer } from 'apollo-server';
+import { typeDefs } from './graphql/typeDefs';
+import { resolvers } from './graphql/resolvers';
 const server = new ApolloServer({
   typeDefs,
   resolvers,
